攻撃等制作
sword,gun　技３づつ制作　IDamageの受け渡し調整　IDamage取得のnull検知　Enemyspawn調整

+++ MechaAction/Assets/okamoto/Script/delete/IDamage.cs
@@ -6,6 +6,6 @@ public interface IDamage
 {
     void TakeDamage(int amount,int knockback,int dir);
 
-    void Heal(float amount);
+    void Heal(int amount);
     void Die();
 }
